+++ library/src/main/java/io/sfrei/tracksearch/clients/soundcloud/SoundCloudClient.java
@@ -11,6 +11,7 @@ import io.sfrei.tracksearch.tracks.BaseTrackList;
 import io.sfrei.tracksearch.tracks.SoundCloudTrack;
 import io.sfrei.tracksearch.tracks.Track;
 import io.sfrei.tracksearch.tracks.TrackList;
+import io.sfrei.tracksearch.utils.ExceptionUtility;
 import io.sfrei.tracksearch.utils.TrackListHelper;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
@@ -57,33 +58,36 @@ public class SoundCloudClient extends SingleSearchClient<SoundCloudTrack> {
         return new HashMap<>(Map.of(TrackList.QUERY_PARAM, query));
     }
 
-    private BaseTrackList<SoundCloudTrack> getTracksForSearch(final String search, int position, int offset) throws TrackSearchException {
+    private BaseTrackList<SoundCloudTrack> getTracksForSearch(final String search, int position, int offset, QueryType queryType)
+            throws TrackSearchException {
+
         checkClientIDAvailableOrThrow();
 
         final Map<String, String> pagingParams = getPagingParams(position, offset);
         final ResponseWrapper response = getSearch(search, true, pagingParams);
 
         final String content = response.getContentOrThrow();
-        return soundCloudUtility.getSoundCloudTracks(content, QueryType.SEARCH, search, this::provideStreamUrl);
+        return soundCloudUtility.getSoundCloudTracks(content, queryType, search, this::provideStreamUrl);
     }
 
     @Override
     public TrackList<SoundCloudTrack> getNext(@NonNull final TrackList<? extends Track> trackList) throws TrackSearchException {
         throwIfPagingValueMissing(this, trackList);
 
-        if (trackList.getQueryType().equals(QueryType.SEARCH)) {
+        final QueryType trackListQueryType = trackList.getQueryType();
+        if (trackListQueryType.equals(QueryType.SEARCH) || trackListQueryType.equals(QueryType.PAGING)) {
             final int queryPosition = trackList.getQueryInformationIntValue(OFFSET_KEY);
             final int queryOffset = TrackSearchConfig.DEFAULT_PLAYLIST_OFFSET;
 
-            final BaseTrackList<SoundCloudTrack> nextTracksForSearch = getTracksForSearch(trackList.getQueryParam(), queryPosition, queryOffset);
+            final BaseTrackList<SoundCloudTrack> nextTracksForSearch = getTracksForSearch(trackList.getQueryParam(), queryPosition, queryOffset, QueryType.PAGING);
             return TrackListHelper.updatePagingValues(nextTracksForSearch, trackList, POSITION_KEY, OFFSET_KEY);
         }
-        throw new SoundCloudException("Query type not supported");
+        throw new SoundCloudException(ExceptionUtility.unsupportedQueryTypeMessage(trackListQueryType));
     }
 
     private String provideStreamUrl(final SoundCloudTrack track) {
         try {
-            return getStreamUrl(track, TrackSearchConstants.RETRY_RESOLVING_ONCE);
+            return getStreamUrl(track, TrackSearchConstants.DEFAULT_RESOLVING_RETRIES);
         } catch (TrackSearchException e) {
             log.error("Error occurred acquiring stream URL", e);
         }
@@ -106,7 +110,7 @@ public class SoundCloudClient extends SingleSearchClient<SoundCloudTrack> {
     @Override
     public String getStreamUrl(@NonNull SoundCloudTrack soundCloudTrack, final int retries) throws TrackSearchException {
         return ClientHelper.getStreamUrl(this, soundCloudTrack, this::requestAndGetCode, retries)
-                .orElseThrow(() -> new SoundCloudException(String.format("Not able to get stream URL after %s retries", retries)));
+                .orElseThrow(() -> new SoundCloudException(ExceptionUtility.noStreamUrlAfterRetriesMessage(retries)));
     }
 
     private ResponseWrapper getSearch(final String search, final boolean firstRequest, final Map<String, String> pagingParams) {
@@ -166,7 +170,7 @@ public class SoundCloudClient extends SingleSearchClient<SoundCloudTrack> {
 
     @Override
     public TrackList<SoundCloudTrack> getTracksForSearch(@NonNull final String search) throws TrackSearchException {
-        BaseTrackList<SoundCloudTrack> trackList = getTracksForSearch(search, 0, TrackSearchConfig.DEFAULT_PLAYLIST_OFFSET);
+        BaseTrackList<SoundCloudTrack> trackList = getTracksForSearch(search, 0, TrackSearchConfig.DEFAULT_PLAYLIST_OFFSET, QueryType.SEARCH);
         trackList.addQueryInformationValue(POSITION_KEY, 0);
         return trackList;
     }