+++ library/src/main/java/io/sfrei/tracksearch/config/TrackSearchConstants.java
@@ -2,6 +2,6 @@ package io.sfrei.tracksearch.config;
 
 public interface TrackSearchConstants {
 
-    int RETRY_RESOLVING_ONCE = 1;
+    int DEFAULT_RESOLVING_RETRIES = 2;
 
 }