+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 Changelog
 =========
 
+0.7.1
+-----
+
+**Feature:**
+
+- Additional youtube video data path
+- Increase default URL resolving retries
+- Avoid resolving of streams and ads
+- Updated dependencies
+
+**Bugfix:**
+  
+- Fix youtube track length resolving
+
 0.7.0
 -----